Move bitcoin serialisers to v1 package
