+++ src/commonMain/kotlin/fr/acinq/lightning/transactions/Transactions.kt
@@ -24,6 +24,7 @@ import fr.acinq.lightning.blockchain.fee.FeeratePerKw
 import fr.acinq.lightning.channel.Commitments
 import fr.acinq.lightning.io.*
 import fr.acinq.lightning.serialization.*
+import fr.acinq.lightning.serialization.v1.*
 import fr.acinq.lightning.transactions.CommitmentOutput.InHtlc
 import fr.acinq.lightning.transactions.CommitmentOutput.OutHtlc
 import fr.acinq.lightning.utils.*
